fix: added extra command fee to dpos
